(chore): introduce sentry to ssr

+++ src/app/services/storage.ts
@@ -1,9 +1,16 @@
+import { isPlatformServer } from '@angular/common';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
+
+@Injectable()
 export class Storage {
-  static _() {
-    return new Storage();
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
+
+  static _(platformId: Object) {
+    return new Storage(platformId);
   }
 
   get(key: string) {
+    if (isPlatformServer(this.platformId)) return;
     try {
       return window.localStorage.getItem(key);
     } catch (err) {
@@ -14,13 +21,16 @@ export class Storage {
     }
   }
   set(key: string, value: any) {
+    if (isPlatformServer(this.platformId)) return;
     return window.localStorage.setItem(key, value);
   }
   destroy(key: string) {
+    if (isPlatformServer(this.platformId)) return;
     return window.localStorage.removeItem(key);
   }
 
   clear() {
+    if (isPlatformServer(this.platformId)) return;
     window.localStorage.clear();
   }
 }
